pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integ"
version = "0.1.0"
description = "Linux platform for GOG & Steam game integration with Proton support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0" }
authors = [
    { name = "rheldev6", email = "rheldev6@gmail.com" },
]
keywords = ["gog", "steam", "gaming", "proton", "linux"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

dependencies = [
    # Frontend
    "PyQt6>=6.6.0",
    "PyQt6-sip>=13.6.0",
    
    # Backend
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    
    # Utils
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

backend = [
    "celery>=5.3.0",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
]

[tool.setuptools]
packages = ["frontend", "backend"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
